requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
psycopg2>=2.8
redis>=3.5
celery>=5.0
django-environ==0.11.2
djangorestframework==3.15.1
